Provide the same CAPA response in both pop3front-auth and -maildir.
The RFC 2449 standard states that the CAPA response MUST be effectively
identical in both AUTHORIZATION and TRNANSACTION states "to check for
active down-negotiation attacks."

This unfortunately causes pop3front-maildir to be linked against the CVM
libraries just to grab the SASL parameters.  I may yet rewrite this
change to use environment variables to avoid this extraneous linkage.