imapfront-auth: Add support for running Dovecot's IMAP process
The Dovecot imap program requires $MAIL to be set specially.  The CVM
library already sets $MAIL to the path to the mailbox/maildir, but
Dovecot expects two modifications:

1. The path must be prefixed with either "maildir:" or "mbox:"

2. Any instances of ":" within the path must be replaced by "::".

This change does both of the above if $SETUP_ENV=dovecot.  It relies on
a change introduced in cvm-0.82 to set $MAIL and $MAILDIR to
cvm_fact_mailbox.