Eliminate the short-circuit on accept from plugins
Previously, when a plugin returned an "accept" code from the sender or
recipient hook, the core would not call any further plugins in the
chain.  This caused a problem for the new counters plugin handling the
SMTP SIZE parameter.

With this change, the first "accept" code is memorized, but the rest of
the plugins are still run, allowing later plugins to still reject the
sender or recipient.